+++ app/src/main/java/org/horaapps/leafpic/SelectAlbumBottomSheet.java
@@ -27,8 +27,8 @@ import com.mikepenz.community_material_typeface_library.CommunityMaterial;
 import com.mikepenz.iconics.view.IconicsImageView;
 
 import org.horaapps.leafpic.activities.base.ThemedActivity;
-import org.horaapps.leafpic.data.Album;
-import org.horaapps.leafpic.data.base.FoldersFileFilter;
+import org.horaapps.leafpic.model.Album;
+import org.horaapps.leafpic.model.base.FoldersFileFilter;
 import org.horaapps.leafpic.util.AlertDialogsHelper;
 import org.horaapps.leafpic.util.ThemeHelper;
