+++ app/src/main/java/org/horaapps/leafpic/activities/MainActivity.java
@@ -53,13 +53,13 @@ import org.horaapps.leafpic.SelectAlbumBottomSheet;
 import org.horaapps.leafpic.activities.base.SharedMediaActivity;
 import org.horaapps.leafpic.adapters.AlbumsAdapter;
 import org.horaapps.leafpic.adapters.MediaAdapter;
-import org.horaapps.leafpic.data.Album;
-import org.horaapps.leafpic.data.CustomAlbumsHelper;
-import org.horaapps.leafpic.data.HandlingAlbums;
-import org.horaapps.leafpic.data.Media;
-import org.horaapps.leafpic.data.base.FilterMode;
-import org.horaapps.leafpic.data.base.SortingMode;
-import org.horaapps.leafpic.data.base.SortingOrder;
+import org.horaapps.leafpic.model.Album;
+import org.horaapps.leafpic.model.CustomAlbumsHelper;
+import org.horaapps.leafpic.model.HandlingAlbums;
+import org.horaapps.leafpic.model.Media;
+import org.horaapps.leafpic.model.base.FilterMode;
+import org.horaapps.leafpic.model.base.SortingMode;
+import org.horaapps.leafpic.model.base.SortingOrder;
 import org.horaapps.leafpic.util.Affix;
 import org.horaapps.leafpic.util.AlertDialogsHelper;
 import org.horaapps.leafpic.util.ContentHelper;
