+++ app/src/main/java/org/horaapps/leafpic/activities/SplashScreen.java
@@ -16,7 +16,7 @@ import android.widget.Toast;
 
 import org.horaapps.leafpic.R;
 import org.horaapps.leafpic.activities.base.SharedMediaActivity;
-import org.horaapps.leafpic.data.Album;
+import org.horaapps.leafpic.model.Album;
 import org.horaapps.leafpic.util.ColorPalette;
 import org.horaapps.leafpic.util.PermissionUtils;
 import org.horaapps.leafpic.util.PreferenceUtil;
