+++ app/src/main/java/org/horaapps/leafpic/activities/base/SharedMediaActivity.java
@@ -3,8 +3,8 @@ package org.horaapps.leafpic.activities.base;
 import android.os.Bundle;
 
 import org.horaapps.leafpic.MyApplication;
-import org.horaapps.leafpic.data.Album;
-import org.horaapps.leafpic.data.HandlingAlbums;
+import org.horaapps.leafpic.model.Album;
+import org.horaapps.leafpic.model.HandlingAlbums;
 
 /**
  * Created by dnld on 03/08/16.
