+++ app/src/main/java/org/horaapps/leafpic/adapters/MediaAdapter.java
@@ -17,7 +17,7 @@ import com.koushikdutta.ion.Ion;
 import com.mikepenz.community_material_typeface_library.CommunityMaterial;
 import com.mikepenz.iconics.view.IconicsImageView;
 
-import org.horaapps.leafpic.data.Media;
+import org.horaapps.leafpic.model.Media;
 import org.horaapps.leafpic.util.ColorPalette;
 import org.horaapps.leafpic.util.ThemeHelper;
