+++ app/src/main/java/org/horaapps/leafpic/adapters/MediaPagerAdapter.java
@@ -8,7 +8,7 @@ import android.util.SparseArray;
 import android.view.View;
 import android.view.ViewGroup;
 
-import org.horaapps.leafpic.data.Media;
+import org.horaapps.leafpic.model.Media;
 import org.horaapps.leafpic.fragments.GifFragment;
 import org.horaapps.leafpic.fragments.ImageFragment;
 import org.horaapps.leafpic.fragments.VideoFragment;
