+++ app/src/main/java/org/horaapps/leafpic/model/SimpleMediaIdentifier.java
@@ -1,4 +1,4 @@
-package org.horaapps.leafpic.data;
+package org.horaapps.leafpic.model;
 
 /**
  * Created by brand on 8/12/2016.
