+++ app/src/main/java/org/horaapps/leafpic/model/base/FoldersFileFilter.java
@@ -1,4 +1,4 @@
-package org.horaapps.leafpic.data.base;
+package org.horaapps.leafpic.model.base;
 
 import java.io.File;
 import java.io.FileFilter;
