+++ app/src/main/java/org/horaapps/leafpic/model/providers/StorageProvider.java
@@ -1,12 +1,12 @@
-package org.horaapps.leafpic.data.providers;
+package org.horaapps.leafpic.model.providers;
 
 import android.content.Context;
 
-import org.horaapps.leafpic.data.Album;
-import org.horaapps.leafpic.data.CustomAlbumsHelper;
-import org.horaapps.leafpic.data.Media;
-import org.horaapps.leafpic.data.base.FoldersFileFilter;
-import org.horaapps.leafpic.data.base.ImageFileFilter;
+import org.horaapps.leafpic.model.Album;
+import org.horaapps.leafpic.model.CustomAlbumsHelper;
+import org.horaapps.leafpic.model.Media;
+import org.horaapps.leafpic.model.base.FoldersFileFilter;
+import org.horaapps.leafpic.model.base.ImageFileFilter;
 import org.horaapps.leafpic.util.ContentHelper;
 import org.horaapps.leafpic.util.PreferenceUtil;
