+++ build.gradle
@@ -6,7 +6,7 @@ buildscript {
         maven { url "https://plugins.gradle.org/m2/" }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.2.1'
+        classpath 'com.android.tools.build:gradle:2.2.2'
         classpath "gradle.plugin.com.gladed.gradle.androidgitversion:gradle-android-git-version:0.2.2"
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
